+++ src/sphinx_substitution_extensions/shared.py
@@ -6,3 +6,6 @@ Constants and functions shared between modules.
 # may break doc8 linting.
 # See https://github.com/PyCQA/doc8/pull/34.
 SUBSTITUTION_OPTION_NAME = "substitutions"
+
+CONTENT_SUBSTITUTION_OPTION_NAME = "content-substitutions"
+PATH_SUBSTITUTION_OPTION_NAME = "path-substitutions"